--- packages/eas-cli/src/credentials/ios/actions/AppleAppIdentifierUtils.ts
@@ -1,4 +1,0 @@
-export function isWildcardBundleIdentifier(bundleIdentifier: string): boolean {
-  const wildcardRegex = /^[A-Za-z0-9.-]+\*$/;
-  return wildcardRegex.test(bundleIdentifier);
-}
