--- packages/eas-cli/src/credentials/ios/actions/__tests__/AppleAppIdentifierUtils-test.ts
@@ -1,13 +1,0 @@
-import { isWildcardBundleIdentifier } from '../AppleAppIdentifierUtils';
-
-describe(isWildcardBundleIdentifier, () => {
-  it('classifies wildcard bundle identifiers correctly', async () => {
-    expect(isWildcardBundleIdentifier('doge.doge.*')).toBe(true);
-    expect(isWildcardBundleIdentifier('doge*')).toBe(true);
-
-    expect(isWildcardBundleIdentifier('*')).toBe(false);
-    expect(isWildcardBundleIdentifier('*.doge')).toBe(false);
-    expect(isWildcardBundleIdentifier('doge')).toBe(false);
-    expect(isWildcardBundleIdentifier('doge.doge.doge')).toBe(false);
-  });
-});
