+++ packages/eas-cli/src/project/ios/__tests__/bundleIdentifier-test.ts
@@ -8,6 +8,7 @@ import { promptAsync } from '../../../prompts';
 import {
   ensureBundleIdentifierIsDefinedForManagedProjectAsync,
   getBundleIdentifier,
+  isWildcardBundleIdentifier,
 } from '../bundleIdentifier';
 
 jest.mock('fs');
@@ -139,5 +140,17 @@ describe(ensureBundleIdentifierIsDefinedForManagedProjectAsync, () => {
         expo: { ios: { bundleIdentifier: 'com.expo.notdominik' } },
       });
     });
+  });
+});
+
+describe(isWildcardBundleIdentifier, () => {
+  it('classifies wildcard bundle identifiers correctly', async () => {
+    expect(isWildcardBundleIdentifier('doge.doge.*')).toBe(true);
+    expect(isWildcardBundleIdentifier('doge*')).toBe(true);
+
+    expect(isWildcardBundleIdentifier('*')).toBe(false);
+    expect(isWildcardBundleIdentifier('*.doge')).toBe(false);
+    expect(isWildcardBundleIdentifier('doge')).toBe(false);
+    expect(isWildcardBundleIdentifier('doge.doge.doge')).toBe(false);
   });
 });
