+++ packages/eas-cli/src/project/ios/bundleIdentifier.ts
@@ -130,3 +130,8 @@ function _warnIfBundleIdentifierDefinedInAppConfigForGenericProject(
 export const warnIfBundleIdentifierDefinedInAppConfigForGenericProject = once(
   _warnIfBundleIdentifierDefinedInAppConfigForGenericProject
 );
+
+export function isWildcardBundleIdentifier(bundleIdentifier: string): boolean {
+  const wildcardRegex = /^[A-Za-z0-9.-]+\*$/;
+  return wildcardRegex.test(bundleIdentifier);
+}
